// packages/receipt/src/generated/types.ts
export interface BasicHeaderType {
  /**
   * Identifier of the request/response, unique with the taxnumber in every data
   * exchange transaction
   *
   * A kérés/válasz azonosítója, minden üzenetváltásnál - adószámonként - egyedi
   */
  requestId: string;
  /**
   * UTC time of the request/response
   *
   * A kérés/válasz keletkezésének UTC ideje
   */
  timestamp: string;
  /**
   * Request version number, indicating which datastructure the client sends data
   * in, and in which the response is expected
   *
   * A kérés/válasz verziószáma, hogy a hívó melyik interfész verzió szerint küld
   * adatot és várja a választ
   */
  requestVersion: string;
  /**
   * Header version number
   *
   * A header verziószáma
   */
  headerVersion?: string;
}
// ...
export interface BasicResultType {
  /**
   * Processing result
   *
   * Feldolgozási eredmény
   */
  funcCode: FunctionCodeType;
  /**
   * Processing error code
   *
   * A feldolgozási hibakód
   */
  errorCode?: string;
  /**
   * Processing message
   *
   * Feldolgozási üzenet
   */
  message?: string;
  /**
   * Miscellaneous notifications
   *
   * Egyéb értesítések
   */
  notifications?: NotificationsType;
}
// ...
export type CertificateTypeType = 'AUTHENTICATION' | 'SIGNING';
// ...
export interface EndpointType {
  /**
   * The order in which the endpoint is requested, starting from 1. The endpoint
   * with the lower sequence number should be requested first.
   *
   * Az endpoint hívásának sorrendje, 1-től kezdődően. A kisebb sorszámú
   * endpointtal kell először meghívni.
   */
  sequence: number;
  /**
   * The method of endpoint call.
   *
   * Az endpoint hívásának metódusa.
   */
  httpMethod: HttpMethodType;
  /**
   * The name of the service which has to be called by the e-cash register.
   *
   * A szolgáltatás neve, amelyet meg kell hívjon az e-pénztárgép.
   */
  service: ServiceType;
  /**
   * URL of the requested service
   *
   * A meghívandó szolgáltatás URL-je
   */
  url: string;
}
// ...
export type FunctionCodeType = 'OK' | 'ERROR';
// ...
export type HttpMethodType = 'GET' | 'POST';
// ...
export interface NotificationsType {
  /**
   * Notification
   *
   * Értesítés
   */
  notification: NotificationType[];
}

/**
 * Notification
 *
 * Értesítés
 */
export interface NotificationType {
  /**
   * Notification code
   *
   * Értesítés kód
   */
  notificationCode: string;
  /**
   * Notification text
   *
   * Értesítés szöveg
   */
  notificationText: string;
}

/**
 * e-Cash register taxpayer and site location details type
 *
 * e-Pénztárgép adózói és üzemeltetési hely adatai típus
 */
export interface OperatorSiteType {
  /**
   * e-Cash register taxpayer and site location details start date of validation
   *
   * e-Pénztárgép adózói és üzemeltetési hely adatai érvényesség kezdete
   */
  operatorSiteValidFrom: string;
  /**
   * e-Cash register location of operation
   *
   * Az e-pénztárgép üzemeltetési helye
   */
  shop: ShopType;
  /**
   * e-Cash register taxpayer data, in case of a fuel owner two-operator e-cash
   * register used on a fuel well
   *
   * e-Pénztárgép üzemeltető adatai, üzemanyagkúton használt kétvállalkozós
   * e-pénztárgép esetén az üzemanyag tulajdonos adózó adatai
   */
  taxPayer: TaxPayerType;
  /**
   * In case of a fuel owner two-operator e-cash register used on a fuel well
   *
   * Üzemanyagkúton használt kétvállalkozós e-pénztárgépen az üzemeltető adózó
   * adatai
   */
  coTaxPayer?: TaxPayerType;
}
// ...
export interface RegistrationResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * e-Cash register AP number
   *
   * e-Pénztárgép AP száma
   */
  APNumber: string;
  /**
   * Callback required
   *
   * Visszahívás szükséges
   */
  callbackRequired: boolean;
  /**
   * Endpoint for downloading the authentication certificate.
   *
   * Az autentikációs tanúsítvány letöltésére való endpoint.
   */
  authenticationCertificateEndpoint?: EndpointType;
  /**
   * Endpoint for downloading the signing certificate.
   *
   * Az aláíró tanúsítvány letöltésére való endpoint.
   */
  signingCertificateEndpoint: EndpointType;
  /**
   * The VAT key and collector list used by e-cash register.
   *
   * Az e-pénztárgépen alkalmazandó ÁFA kulcsok és gyűjtők listája.
   */
  vat: VatType;
  /**
   * The operator and site data of the e-cash register.
   *
   * Az e-pénztárgép üzemeltetői és üzemeltetési hely adatai.
   */
  operatorSite: OperatorSiteType;
  /**
   * EPD process identifier of the VAT table setting instruction to be executed.
   * This identifier must be returned in the currentVatProcessId field in the
   * status report before the hello call.
   *
   * Végrehajtandó ÁFA táblázat beállítás utasítás EPD folyamat azonosítója. A
   * hello hívás előtti állapot visszajelentésben ezt az azonosítót kell
   * visszaadni a currentVatProcessId mezőben.
   */
  vatProcessId: string;
  /**
   * EPD process identifier of the operator and site data setting instruction to
   * be executed. This identifier must be returned in the
   * currentOperatorSiteProcessId field in the status report before the hello
   * call.
   *
   * Végrehajtandó adózói és üzemeltetési hely adatok beállítása utasítás EPD
   * folyamat azonosítója. A hello hívás előtti állapot visszajelentésben ezt az
   * azonosítót kell visszaadni a currentOperatorSiteProcessId mezőben.
   */
  operatorSiteProcessId: string;
}
// ...
export interface RenewCertificateResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * e-Cash register AP number
   *
   * e-Pénztárgép AP száma
   */
  APNumber: string;
  /**
   * Callback required
   *
   * Visszahívás szükséges
   */
  callbackRequired: boolean;
  /**
   * Type of certificate.
   *
   * A tanúsítvány típusa.
   */
  certificateType: CertificateTypeType;
  /**
   * Endpoint for downloading the certificate.
   *
   * A tanúsítvány letöltésére való endpoint.
   */
  certificateEndpoint: EndpointType;
}
// ...
export type ServiceType =
  | 'BLOCK_UNBLOCK'
  | 'DOWNLOAD_PRODUCT_LIST'
  | 'DOWNLOAD_SOFTWARE_UPDATE'
  | 'END_OF_OPERATION'
  | 'QUERY_CERTIFICATE'
  | 'HELLO'
  | 'OPERATOR_SITE_UPDATE'
  | 'OWNER_CHANGE'
  | 'PRINT_TECHNICAL_INFO'
  | 'QUERY_TAXPAYER'
  | 'REGISTRATION'
  | 'RENEW_CERTIFICATE'
  | 'RENEW_EXPIRED_CERTIFICATE'
  | 'SOFTWARE_UPDATE'
  | 'VAT_UPDATE'
  | 'SOUND_DOWNLOAD'
  | 'LOGO_DOWNLOAD'
  | 'CASH_REGISTER_INFO'
  | 'SEND_MISSING_DOCUMENT';

/**
 * Shipping dates
 *
 * Szállítási dátumok
 */
export interface ShippingDatesType {
  /**
   * Shipping date
   *
   * Szállítási dátum
   */
  shippingDate: string[];
}

/**
 * Shop data type
 *
 * Üzlet adatai típus
 *
 * @remarks Exactly one of `simpleAddress`, `detailedAddress` may be set.
 */
export interface ShopType {
  /**
   * Simple address
   *
   * Egyszerű cím
   */
  simpleAddress?: EreceiptBaseSimpleAddressType;
  /**
   * Detailed address
   *
   * Részletes cím
   */
  detailedAddress?: EreceiptBaseDetailedAddressType;
  /**
   * Shop name
   *
   * Üzlet neve
   */
  shopName: string;
  /**
   * Shop short name
   *
   * Üzlet rövid neve
   */
  shopShortName?: string;
}
// ...
export type SoftwareOperationType =
  'AE' | 'ONLINE_ANDROID' | 'ONLINE_IOS' | 'ONLINE_HUAWEI' | 'ONLINE_SERVICE';

/**
 * Software data type
 *
 * A program adatai típus
 */
export interface SoftwareType {
  /**
   * Software ID
   *
   * A program azonosítója
   */
  softwareId: string;
  /**
   * Software name
   *
   * A program neve
   */
  softwareName: string;
  /**
   * Software operation method (AE-based (hardware) or online)
   *
   * A program működési típusa (AE alapú (hardveres) vagy online szolgáltatás)
   */
  softwareOperation: SoftwareOperationType;
  /**
   * Software main version
   *
   * A program főverziója
   */
  softwareMainVersion: string;
  /**
   * The verification code of the software calculated with the SHA256 algorithm
   *
   * A szoftver SHA256 algoritmussal kiszámított ellenőrző kódja
   */
  softwareHash: string;
  /**
   * Software last update time
   *
   * A szoftver legutóbbi frissítésének időpontja
   */
  softwareLastUpdateTime: string;
  /**
   * Name of the software's developer
   *
   * A program fejlesztőjének neve
   */
  softwareDevName: string;
  /**
   * Electronic contact of the software's developer
   *
   * A program fejlesztőjének elektronikus elérhetősége
   */
  softwareDevContact: string;
  /**
   * ISO-3166 alpha2 country code of the software's developer
   *
   * A program fejlesztőjének ISO-3166 alpha2 országkódja
   */
  softwareDevCountryCode?: string;
  /**
   * Tax number of the software's developer
   *
   * A program fejlesztőjének adószáma
   */
  softwareDevTaxNumber?: string;
}
// ...
export interface TaxPayerType {
  /**
   * Simple address
   *
   * Egyszerű cím
   */
  simpleAddress?: EreceiptBaseSimpleAddressType;
  /**
   * Detailed address
   *
   * Részletes cím
   */
  detailedAddress?: EreceiptBaseDetailedAddressType;
  /**
   * Tax payer name
   *
   * Adózó neve
   */
  taxPayerName: string;
  /**
   * Tax payer short name
   *
   * Adózó rövid neve
   */
  taxPayerShortName?: string;
  /**
   * Tax payer tax number or member tax number if part of a VAT group.
   *
   * Adózó adószáma, vagy ÁFA-csoport tagság esetén a tagi adószáma.
   */
  taxNumber: EreceiptBaseTaxNumberType;
  /**
   * Group identification number if part of a VAT group
   *
   * ÁFA-csoport tagság esetén a csoportazonosító szám
   */
  groupIdentificationNumber?: EreceiptBaseTaxNumberType;
}
// ...
export interface VatGroupType {
  /**
   * Collector code on 3 characters
   *
   * Forgalmi gyűjtő jele 3 karakteren
   */
  collectorCode: string;
  /**
   * The valid VAT rate, expressed as a percentage, without a percentage sign
   * and, where necessary, with a decimal point
   *
   * Az érvényes ÁFA-kulcs, százalékban kifejezve, százalékjel nélkül, szükség
   * esetén tizedespontot alkalmazva
   */
  vatPercentage: string;
  /**
   * The VAT content of VAT rate, expressed as a percentage, without a percentage
   * sign and, where necessary, with a decimal point
   *
   * Az adott ÁFA kulcshoz tartozó ÁFA tartalom, százalékban kifejezve,
   * százalékjel nélkül, szükség esetén tizedespontot alkalmazva
   */
  vatContent: string;
  /**
   * The rate of each VAT rate expressed as a whole number, with percentage sign
   *
   * Az egyes ÁFA kulcsok mértéke egész számban kifejezve, százalékjellel
   */
  vatLabel: string;
}
// ...
export interface VatType {
  /**
   * Validation date, after the VAT rates and category labels must be applied on
   * the e-cash register
   *
   * Azon időpont, amit követően az ÁFA kulcsokat és kategória címkéket
   * alkalmazni kell az e-pénztárgépen
   */
  vatValidFrom: string;
  /**
   * Data of the VAT rate
   *
   * Adott ÁFA kulcs adatai
   *
   * @remarks At most 16 entries.
   */
  vatGroup: VatGroupType[];
}